Add section on nix language.
This covers basics about the nix language, but then links to the
manual.  Do we want to have any references to syntax in our guide, or
just link straight to the nix manual?

Fundamentally the question is whether the target demographic of this
guide is potentially different than the target demographic of the nix
manuals.