Side notes for users not seeing matching output
For readers who have not used the Determinate Nix Installer, they'll probably
need to activate the nix command and flakes, and they may not get the change
in command prompt in entry 3. This is the case if someone is using the official
installer or trying to read through the guide on nixos. Closes #315.